🚚 Move Category-related files to a separated packaged
Now that this component and UI state will be reused, it is better to
move it to a dedicated package.
