+++ features/task/src/androidTest/java/com/escodro/task/CategorySelectionTest.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.task.model.Category
-import com.escodro.task.presentation.detail.category.CategorySelection
-import com.escodro.task.presentation.detail.category.ChipNameKey
-import com.escodro.task.presentation.detail.category.TaskCategoryState
+import com.escodro.task.presentation.category.CategorySelection
+import com.escodro.task.presentation.category.ChipNameKey
+import com.escodro.task.presentation.category.CategoryState
 import com.escodro.theme.AlkaaTheme
 import org.junit.Ignore
 import org.junit.Rule
@@ -110,7 +110,7 @@ internal class CategorySelectionTest {
         composeTestRule.setContent {
             AlkaaTheme {
                 CategorySelection(
-                    state = TaskCategoryState.Empty,
+                    state = CategoryState.Empty,
                     currentCategory = null,
                     onCategoryChanged = { }
                 )
@@ -126,7 +126,7 @@ internal class CategorySelectionTest {
         composeTestRule.setContent {
             AlkaaTheme {
                 CategorySelection(
-                    state = TaskCategoryState.Loaded(categories),
+                    state = CategoryState.Loaded(categories),
                     currentCategory = currentCategory,
                     onCategoryChanged = { }
                 )
