+++ features/task/src/androidTest/java/com/escodro/task/TaskDetailTest.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.task.model.Task
 import com.escodro.task.presentation.detail.TaskDetailActions
-import com.escodro.task.presentation.detail.category.TaskCategoryState
+import com.escodro.task.presentation.category.CategoryState
 import com.escodro.task.presentation.detail.main.TaskDetailRouter
 import com.escodro.task.presentation.detail.main.TaskDetailState
 import com.escodro.theme.AlkaaTheme
@@ -59,7 +59,7 @@ internal class TaskDetailTest {
             AlkaaTheme {
                 TaskDetailRouter(
                     detailViewState = state,
-                    categoryViewState = TaskCategoryState.Loaded(listOf()),
+                    categoryViewState = CategoryState.Loaded(listOf()),
                     actions = TaskDetailActions()
                 )
             }
