--- features/task/src/main/java/com/escodro/task/presentation/detail/category/TaskCategoryState.kt
@@ -1,10 +1,0 @@
-package com.escodro.task.presentation.detail.category
-
-import com.escodro.task.model.Category
-
-internal sealed class TaskCategoryState {
-
-    data class Loaded(val categoryList: List<Category>) : TaskCategoryState()
-
-    object Empty : TaskCategoryState()
-}
